Removed compiler warnings:
./include/pvfs2-types.h: In function ‘encode_PVFS_credential’:
./include/pvfs2-types.h:1213: warning: comparison between signed and unsigned integer expressions



SVN revision 11702

+++ src/proto/endecode-funcs.h
@@ -958,14 +958,14 @@ static inline void encode_##name(char **pptr, const struct name *x) { int i; \
     encode_##t3(pptr, &x->x3); \
     encode_##tn1(pptr, &x->n1); \
     if (x->n1 > 0) \
-        for (i=0; i<x->n1; i++) \
+        for (i=0; i<(int)(x->n1); i++) \
             encode_##ta1(pptr, &(x)->a1[i]); \
     align8(pptr); \
     encode_##t4(pptr, &x->x4); \
     encode_##t5(pptr, &x->x5); \
     encode_##tn2(pptr, &x->n2); \
     if (x->n2 > 0) \
-        for (i=0; i<x->n2; i++) \
+        for (i=0; i<(int)(x->n2); i++) \
             encode_##ta2(pptr, &(x)->a2[i]); \
     align8(pptr); \
     encode_##t6(pptr, &x->x6); \
@@ -979,7 +979,7 @@ static inline void decode_##name(char **pptr, struct name *x) { int i; \
     if (x->n1 > 0) \
     { \
         x->a1 = decode_malloc(x->n1 * sizeof(*x->a1)); \
-        for (i=0; i<x->n1; i++) \
+        for (i=0; i<(int)(x->n1); i++) \
             decode_##ta1(pptr, &(x)->a1[i]); \
     } \
     else \
@@ -991,7 +991,7 @@ static inline void decode_##name(char **pptr, struct name *x) { int i; \
     if (x->n2 > 0) \
     { \
         x->a2 = decode_malloc(x->n2 * sizeof(*x->a2)); \
-        for (i=0; i<x->n2; i++) \
+        for (i=0; i<(int)(x->n2); i++) \
             decode_##ta2(pptr, &(x)->a2[i]); \
     } \
     else \
